Puzzle 9: Second part done

+++ src/test/java/fr/syalioune/puzzle/xmas/XmasErrorDetectorShould.java
@@ -39,7 +39,7 @@ public class XmasErrorDetectorShould {
 
   @ParameterizedTest
   @MethodSource("invalidListAndOneInputAfterPreambleSource")
-  public void detectOneErrorWithInvalidListAndOneInputAfterPreamble(List<Double> input, int preambleSize, int badNumber) {
+  public void detectOneErrorWithInvalidListAndOneInputAfterPreamble(List<Double> input, int preambleSize, double badNumber) {
     // Arrange
     XmasErrorDetector xmasErrorDetector = new XmasErrorDetector(preambleSize);
 
@@ -66,7 +66,7 @@ public class XmasErrorDetectorShould {
 
   @ParameterizedTest
   @MethodSource("invalidListAndMoreThanOneOneInputAfterPreambleSource")
-  public void detectOneErrorWithInvalidListAndMoreThanOneInputAfterPreamble(List<Double> input, int preambleSize, int badNumber) {
+  public void detectOneErrorWithInvalidListAndMoreThanOneInputAfterPreamble(List<Double> input, int preambleSize, double badNumber) {
     // Arrange
     XmasErrorDetector xmasErrorDetector = new XmasErrorDetector(preambleSize);
 
@@ -82,31 +82,31 @@ public class XmasErrorDetectorShould {
     return Stream.of(
         Arguments.arguments(null,0),
         Arguments.arguments(Collections.emptyList(), 0),
-        Arguments.arguments(Arrays.asList(1,2,3), 5)
+        Arguments.arguments(Arrays.asList(1.0,2.0,3.0), 5)
     );
   }
 
   static Stream<Arguments> validListAndOneInputAfterPreambleSource() {
     return Stream.of(
-        Arguments.arguments(Arrays.asList(1,2,3,4,5,6), 5)
+        Arguments.arguments(Arrays.asList(1.0,2.0,3.0,4.0,5.0,6.0), 5)
     );
   }
 
   static Stream<Arguments> invalidListAndOneInputAfterPreambleSource() {
     return Stream.of(
-        Arguments.arguments(Arrays.asList(1,2,3,4,5,20), 5, 20)
+        Arguments.arguments(Arrays.asList(1.0,2.0,3.0,4.0,5.0,20.0), 5, 20.0)
     );
   }
 
   static Stream<Arguments> validListAndMoreThanOneInputAfterPreambleSource() {
     return Stream.of(
-        Arguments.arguments(Arrays.asList(1,2,3,4,5,6,11,7), 5)
+        Arguments.arguments(Arrays.asList(1.0,2.0,3.0,4.0,5.0,6.0,11.0,7.0), 5)
     );
   }
 
   static Stream<Arguments> invalidListAndMoreThanOneOneInputAfterPreambleSource() {
     return Stream.of(
-        Arguments.arguments(Arrays.asList(35,20,15,25,47,40,62,55,65,95,102,117,150,182,127,219,299), 5, 127)
+        Arguments.arguments(Arrays.asList(35.0,20.0,15.0,25.0,47.0,40.0,62.0,55.0,65.0,95.0,102.0,117.0,150.0,182.0,127.0,219.0,299.0), 5, 127.0)
     );
   }
